# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Setspace Agency Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /video-projects:
    get:
      operationId: listVideoProjects
      tags: [videos]
      summary: List video projects
      parameters:
        - name: taskId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Video projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VideoProject"
    post:
      operationId: createVideoProject
      tags: [videos]
      summary: Create a video project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVideoProjectBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoProject"
  /video-projects/{projectId}:
    get:
      operationId: getVideoProject
      tags: [videos]
      summary: Get a video project with versions
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Video project with versions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoProjectDetail"
    patch:
      operationId: updateVideoProject
      tags: [videos]
      summary: Update a video project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVideoProjectBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoProject"
    delete:
      operationId: deleteVideoProject
      tags: [videos]
      summary: Delete a video project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /video-projects/{projectId}/versions:
    post:
      operationId: createVideoVersion
      tags: [videos]
      summary: Upload a new version of a video
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVideoVersionBody"
      responses:
        "201":
          description: Created version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoVersion"
  /video-versions/{versionId}/approve:
    post:
      operationId: approveVideoVersion
      tags: [videos]
      summary: Approve a video version
      parameters:
        - name: versionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoVersion"
  /video-versions/{versionId}/request-revision:
    post:
      operationId: requestVideoRevision
      tags: [videos]
      summary: Request revision on a video version
      parameters:
        - name: versionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoVersion"
  /video-versions/{versionId}/comments:
    get:
      operationId: listVideoComments
      tags: [videos]
      summary: List comments for a video version
      parameters:
        - name: versionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VideoComment"
    post:
      operationId: createVideoComment
      tags: [videos]
      summary: Add a comment to a video version
      parameters:
        - name: versionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVideoCommentBody"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoComment"
  /video-comments/{commentId}/resolve:
    post:
      operationId: resolveVideoComment
      tags: [videos]
      summary: Resolve a video comment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resolved comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoComment"
  /video-versions/{versionId}/share-token:
    post:
      operationId: createShareToken
      tags: [videos]
      summary: Generate a client share link for a video version
      parameters:
        - name: versionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Share token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareToken"
  /review/{token}:
    get:
      operationId: getPublicReview
      tags: [videos]
      summary: Get public review data for a client link (no auth required)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public review data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicReviewData"
  /review/{token}/comments:
    post:
      operationId: createPublicComment
      tags: [videos]
      summary: Add a client comment via public share link
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePublicCommentBody"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoComment"
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        profileImage:
          type: string
          nullable: true
        role:
          type: string
        department:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, username, firstName, lastName, role, createdAt]
...
    VideoProject:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        clientName:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        taskId:
          type: integer
          nullable: true
        status:
          type: string
        createdById:
          type: string
        latestVersionId:
          type: integer
          nullable: true
        latestVersionNumber:
          type: integer
          nullable: true
        latestVersionStatus:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, status, createdById, createdAt, updatedAt]
    VideoProjectDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        clientName:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        taskId:
          type: integer
          nullable: true
        status:
          type: string
        createdById:
          type: string
        versions:
          type: array
          items:
            $ref: "#/components/schemas/VideoVersion"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, status, createdById, versions, createdAt, updatedAt]
    CreateVideoProjectBody:
      type: object
      properties:
        title:
          type: string
        clientName:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        taskId:
          type: integer
          nullable: true
      required: [title]
    UpdateVideoProjectBody:
      type: object
      properties:
        title:
          type: string
        clientName:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        status:
          type: string
    VideoVersion:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        versionNumber:
          type: integer
        objectPath:
          type: string
        fileName:
          type: string
        fileSize:
          type: integer
        status:
          type: string
        uploadedById:
          type: string
        uploadedBy:
          $ref: "#/components/schemas/User"
          nullable: true
        shareToken:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, projectId, versionNumber, objectPath, fileName, fileSize, status, uploadedById, createdAt]
    CreateVideoVersionBody:
      type: object
      properties:
        objectPath:
          type: string
        fileName:
          type: string
        fileSize:
          type: integer
      required: [objectPath, fileName, fileSize]
    VideoComment:
      type: object
      properties:
        id:
          type: integer
        versionId:
          type: integer
        authorId:
          type: string
          nullable: true
        authorName:
          type: string
        authorType:
          type: string
        content:
          type: string
        timestampSeconds:
          type: number
          nullable: true
        isResolved:
          type: boolean
        createdAt:
          type: string
      required: [id, versionId, authorName, authorType, content, isResolved, createdAt]
    CreateVideoCommentBody:
      type: object
      properties:
        content:
          type: string
        timestampSeconds:
          type: number
          nullable: true
      required: [content]
    ShareToken:
      type: object
      properties:
        token:
          type: string
        shareUrl:
          type: string
        versionId:
          type: integer
        expiresAt:
          type: string
          nullable: true
      required: [token, shareUrl, versionId]
    PublicReviewData:
      type: object
      properties:
        projectTitle:
          type: string
        clientName:
          type: string
          nullable: true
        versionNumber:
          type: integer
        objectPath:
          type: string
        fileName:
          type: string
        comments:
          type: array
          items:
            $ref: "#/components/schemas/VideoComment"
      required: [projectTitle, versionNumber, objectPath, fileName, comments]
    CreatePublicCommentBody:
      type: object
      properties:
        content:
          type: string
        authorName:
          type: string
        timestampSeconds:
          type: number
          nullable: true
      required: [content, authorName]