# lib/api-zod/src/generated/types/videoVersion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Setspace Agency Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from "./user";

export interface VideoVersion {
  id: number;
  projectId: number;
  versionNumber: number;
  objectPath: string;
  fileName: string;
  fileSize: number;
  status: string;
  uploadedById: string;
  uploadedBy?: User | null;
  shareToken?: string | null;
  createdAt: string;
}